udpated UI

+++ src/gwt/org/bbop/apollo/gwt/client/GroupPanel.java
@@ -11,6 +11,7 @@ import com.google.gwt.user.cellview.client.Column;
 import com.google.gwt.user.cellview.client.ColumnSortEvent;
 import com.google.gwt.user.cellview.client.DataGrid;
 import com.google.gwt.user.cellview.client.TextColumn;
+import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.Composite;
 import com.google.gwt.user.client.ui.HTML;
 import com.google.gwt.user.client.ui.Widget;
@@ -152,8 +153,10 @@ public class GroupPanel extends Composite {
 
     @UiHandler("deleteButton")
     public void deleteGroup(ClickEvent clickEvent){
-        GroupRestService.deleteGroup(selectedGroupInfo);
-        selectionModel.clear();
+        if(Window.confirm("Delete group "+selectedGroupInfo.getName()+"?")){
+            GroupRestService.deleteGroup(selectedGroupInfo);
+            selectionModel.clear();
+        }
     }
 
     @UiHandler("saveButton")
